refactor/메인 액티비티 - top bar, navigation bar, NavigationGraph 분리
