+++ app/src/main/java/org/sopt/at/views/main/MainTopAppBar.kt
@@ -1,0 +1,70 @@
+package org.sopt.at.views.main
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AccountCircle
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import org.sopt.at.R
+import org.sopt.at.views.navigation.Route
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun MainTopAppBar(
+    modifier: Modifier = Modifier,
+    navController : NavController,
+) {
+    TopAppBar(
+        modifier = modifier
+            .wrapContentSize(),
+        title = {
+            Column (
+                modifier = modifier
+                    .fillMaxWidth()
+                    .wrapContentHeight()) {
+                Row(
+                    modifier = modifier.fillMaxWidth(),
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.SpaceBetween
+                ) {
+                    Text(
+                        text = stringResource(R.string.app_name_tving),
+                        fontWeight = FontWeight.Bold,
+                        textAlign = TextAlign.Start,
+                        modifier = modifier.padding(start = 5.dp),
+                    )
+                }
+            }
+        },
+        actions = {
+            IconButton(onClick = {
+                navController.navigate(Route.SEARCH) {
+                    launchSingleTop = true
+                }
+            }
+            ) {
+                Icon(
+                    Icons.Default.AccountCircle,
+                    contentDescription = stringResource(R.string.search)
+                )
+            }
+        }
+    )
+}
